requirements.txt
joblib==1.5.2
numpy==2.3.5
pandas==2.3.3
python-dateutil==2.9.0
pytz==2025.2
scikit-learn==1.7.2
scipy==1.16.3
setuptools==80.9.0
six==1.17.0
threadpoolctl==3.6.0
tzdata==2025.2
wheel==0.45.1
Flask==2.3.2
waitress==2.1.2
pytest
huggingface_hub==1.2.1
llama_cpp_python==0.3.16
ydata-profiling
matplotlib==3.10.0